[FEAT] 전체 좌석 조회 API 구현

+++ podolab-api/src/main/java/com/podolab/api/concert/ConcertRepository.java
@@ -2,5 +2,5 @@ package com.podolab.api.concert;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ConcertRepository extends JpaRepository<Concert, Integer> {
+public interface ConcertRepository extends JpaRepository<Concert, Long> {
 }
